Better naming
